move Synapse tests
